pyproject.toml
[project]
name = "airas-experiment"
version = "0.1.0"
description = "AI Research Experiment Framework"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "wandb>=0.16.0",
    "datasets>=2.16.0",
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "vllm>=0.5.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "omegaconf>=2.3.0",
    "pydantic>=2.5.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
